# Conflicts:
#	Gemfile
#	Gemfile.lock
#	test/test_helper.rb

+++ app/services/page_ref_services.rb
@@ -1,4 +1,3 @@
-# require 'page_ref.rb'
 class PageRefServices
   attr_accessor :page_ref
 
@@ -8,7 +7,7 @@ class PageRefServices
   end
 
   # Eliminate redundancy in the PageRefs by folding two into one
-  def absorb other
+  def absorb other, force=false
     return if page_ref == other
 
     # Take on all the recipes of the other
@@ -27,16 +26,51 @@ class PageRefServices
       rfm.save
     } if other.respond_to? :referments
 
-    (page_ref.attributes.keys -
-        %w{ id type url errcode status dj_id created_at updated_at aliases error_message }).each { |attrname|
-      page_ref.write_attribute(attrname, other.attributes[attrname]) unless page_ref.attributes[attrname].present?
-    }
-    page_ref.status = other.status
+    # An important question is: which url (and associated status attributes) gets used in the product?
+    # The default is to keep the absorber's attributes unless the other is shown to be good. The 'force'
+    # parameter allows this issue to be forced.
+    aliases = (page_ref.aliases + other.aliases + [other.url, page_ref.url]).uniq
+    if force || ((page_ref.http_status != 200) && (other.http_status == 200))
+      hard_attribs = other.attributes.slice *%w{ errcode http_status error_message url }
+      puts "...taking #{hard_attribs} from other"
+      page_ref.assign_attributes hard_attribs
 
-    # Take on all the urls of the other
-    page_ref.aliases |= other.aliases + [other.url] - [page_ref.url]
+      # Soft attributes are copied only if not already set
+      soft_attribs = page_ref.class.mercury_attributes + %w{ link_text }
+      soft_attribs.each { |attrname|
+        unless page_ref.read_attribute(attrname).present?
+          puts "...absorbing #{attrname} = #{other.read_attribute(attrname)}"
+          page_ref.write_attribute attrname, other.read_attribute(attrname)
+        end
+      }
+    end
+    page_ref.aliases = aliases - [page_ref.url]
+
+    other.destroy if other.id # May not have been saved
     page_ref.save
-    other.destroy
+  end
+
+  # Assert a reference to the given URL, linking back to a referent
+  def self.assert_for_referent(uri, tag_or_referent, type=:Definition )
+    pr = "#{type}PageRef".constantize.fetch uri
+    self.new(pr).assert_referent tag_or_referent if pr.errors.empty?
+    pr
+  end
+
+  # Associate this page_ref with the given referent.
+  # NB: had better be a ReferrableReferent or subclass thereof
+  def assert_referent tag_or_referent
+    rft =
+        case tag_or_referent
+          when Tag
+            Referent.express tag_or_referent
+          else
+            tag_or_referent
+        end
+    if rft && !page_ref.referent_ids.include?(rft.id)
+      page_ref.referents << rft
+      page_ref.save
+    end
   end
 
   # Convert a reference, either creating a new PageRef or merging it with another
@@ -108,6 +142,7 @@ class PageRefServices
     nil
   end
 
+  # Convert any relative paths in PageRef urls by resort to aliases
   def self.join_urls what
     (what.to_s+'PageRef').constantize.where(domain: nil).collect { |pr|
       report = "Joining URLS for #{pr.class} ##{pr.id} '#{pr.url}'"
@@ -146,21 +181,23 @@ class PageRefServices
     reports = []
     # Every recipe needs to have a site and a url (page_ref)
     # reports << Recipe.includes(:referent).to_a.collect { |recipe| "Recipe ##{recipe.id} has no referent (ERROR)" unless recipe.referent }.compact
-    reports << Recipe.includes(:page_ref).to_a.collect { |page_ref| "Recipe ##{page_ref.id} has no page_ref (ERROR)" unless page_ref.page_ref }.compact
 
     reports << RecipePageRef.where(url: nil).count.to_s + ' nil urls in RecipePageRefs'
     reports << RecipePageRef.where(site_id: nil).count.to_s + ' nil sites in RecipePageRefs'
     reports << Recipe.includes(:page_ref).collect { |rcp| true if rcp.page_ref.url.blank? }.compact.count.to_s + " nil URLs in recipes"
 
-    bad = RecipePageRef.where.not(status: [0,1,2], http_status: 200)
+    bad = RecipePageRef.includes(:recipes).where.not(status: [0,1,2], http_status: 200)
     unreachables = Recipe.includes(:page_ref, :gleaning).all.collect { |recipe| recipe if recipe.reachable? == false }.compact
+    headless = Recipe.includes(:page_ref).where(page_ref_id: nil)
     # Every PageRef should have at least tried to go out
     reports += [
-        ((RecipePageRef.virgin.count+RecipePageRef.processing.count).to_s + ' Recipe Page Refs need processing'),
-        (RecipePageRef.where(http_status: nil).count.to_s + ' Recipe Page Refs have no HTTP status'),
-        (bad.count.to_s + " bad recipe refs\n\t" + bad.collect { |rpr| "'#{rpr.url}' (#{rpr.id}) -> #{rpr.http_status}"}.sort.join("\n\t")),
-        (unreachables.count.to_s + " recipe(s) with unreachable URL\n\t" + unreachables.collect { |badrecipe| "#{badrecipe.page_ref.url if badrecipe.page_ref} (#{badrecipe.id}) "}.sort.join("\n\t"))
+        ((RecipePageRef.virgin.includes(:recipes).count+RecipePageRef.processing.count).to_s + ' Recipe Page Refs need processing'),
+        (RecipePageRef.includes(:recipes).where(http_status: nil).count.to_s + ' Recipe Page Refs have no HTTP status'),
+        (bad.count.to_s + " bad recipe refs\n\t" + bad.collect { |rpr| "'#{rpr.url}' (#{rpr.id}) (http_status = #{rpr.http_status}) used by recipe(s) #{rpr.recipe_ids} "}.sort.join("\n\t")),
+        (unreachables.count.to_s + " recipe(s) with unreachable URL (bad PageRef and bad gleaning)\n\t" + unreachables.collect { |badrecipe| "#{badrecipe.page_ref.url if badrecipe.page_ref} (#{badrecipe.id}) "}.sort.join("\n\t"))
     ]
+    reports << (["#{headless.count} recipes with no page ref:"] +
+        headless.collect { |recipe| "Recipe ##{recipe.id} '#{recipe.url}'" }).join("\n\t")
     reports
   end
 
@@ -184,7 +221,11 @@ class PageRefServices
   def self.site_reports
     reports = []
     # Every site needs to have a name (referent) and home (page_ref)
-    reports << Site.includes(:referent).to_a.collect { |site| "Site ##{site.id} ('#{site.home}' has no referent (ERROR)" unless site.referent }.compact.sort
+    reports << Site.includes(:referent).to_a.collect { |site|
+      next if site.referent
+      "Site ##{site.id} ('#{site.home}' has no referent (ERROR) #{('but title is '+site.page_ref.title.to_s) if site.page_ref}" +
+      "\n\t...with #{site.definition_page_refs.count} DefinitionPageRef(s) and #{site.recipe_page_refs.count} RecipePageRef(s)"
+    }.compact.sort
     reports << Site.includes(:page_ref).to_a.collect { |site| "Site ##{site.id} ('#{site.home}' has no page_ref (ERROR)" unless site.page_ref }.compact.sort
 
     # Every PageRef should have at least tried to go out
@@ -193,7 +234,9 @@ class PageRefServices
         (SitePageRef.where(url: nil).count.to_s + ' nil site urls'),
         ((SitePageRef.virgin.count+SitePageRef.processing.count).to_s + ' Site Page Refs need processing'),
         (SitePageRef.where(http_status: nil).count.to_s + ' Site Page Refs have no HTTP status'),
-        (badrefids.count.to_s + " bad SitePageRefs: \n\t" + SitePageRef.where(id: badrefids).collect { |spr| "#{spr.id}: '#{spr.url}' -> #{spr.http_status}"}.join("\n\t"))
+        (badrefids.count.to_s + " bad SitePageRefs: \n\t" + SitePageRef.where(id: badrefids).collect { |spr|
+          "#{spr.id}: '#{spr.url}' (for site ##{spr.site_id}) -> #{spr.http_status}"
+        }.join("\n\t"))
     ]
     reports
   end
@@ -202,6 +245,7 @@ class PageRefServices
   def ensure_status force=false
     # Ensure that each PageRef has status and http_status
     # First, check on the url (may lack host and/or scheme due to earlier bug)
+    sentences = []
     while (uri = safe_parse(sanitize_url page_ref.url)) && (uri.host.blank? || uri.scheme.blank?)
       if uri.scheme.blank?
         uri.scheme = 'http'
@@ -213,27 +257,37 @@ class PageRefServices
         break
       end
     end
-    if page_ref.url_changed? && (other = page_ref.class.find_by(url: page_ref.url)) && (other.id != page_ref.id)
+    if page_ref.url_changed? && (extant = page_ref.class.find_by_url page_ref.url) && (extant.id != page_ref.id)
       # Replacement URL is already being serviced by another PageRef
-      PageRefServices.new(other).absorb page_ref
+      PageRefServices.new(extant).absorb page_ref
       return "Destroyed redundant #{page_ref.class} ##{page_ref.id}"
     end
-    if page_ref.http_status && (page_ref.bad? || page_ref.good?) && !force
-      page_ref.save if page_ref.url_changed?
-    else
-      if page_ref.title.present? # We assume that Mercury has done its job
-        return "Status already 200 on 'good' PageRef##{page_ref.id} '#{page_ref.url}'" if page_ref.http_status == 200 && page_ref.good?
-        page_ref.http_status = 200
-        page_ref.good!
-        "Set status on PageRef##{page_ref.id} '#{page_ref.url}': http_status '#{page_ref.http_status}', status '#{page_ref.status}', error '#{page_ref.error_message}"
-      else
-        page_ref.becomes(PageRef).bkg_enqueue priority: 10 # Must be enqueued as a PageRef b/c subclasses aren't recognized by DJ
-        puts "Enqueued #{page_ref.class.to_s} ##{page_ref.id} '#{page_ref.url}' to get status"
-        page_ref.bkg_wait
-        puts "...returned"
-        "Ran #{page_ref.class.to_s} ##{page_ref.id} '#{page_ref.url}' to get status"
-      end
+    if page_ref.title.present? # We assume that Mercury has done its job
+      return "Status already 200 on 'good' PageRef##{page_ref.id} '#{page_ref.url}'" if page_ref.http_status == 200 && page_ref.good?
+      page_ref.http_status = 200
+      page_ref.good!
+      sentences << "Set status on PageRef##{page_ref.id} '#{page_ref.url}': http_status '#{page_ref.http_status}', status '#{page_ref.status}', error '#{page_ref.error_message}"
     end
+    # Many circumstances under which we go out to check the reference, not nec. for the first time
+    if (page_ref.http_status != 200) || !(page_ref.bad? || page_ref.good?) || page_ref.url_changed? || force
+      page_ref.becomes(PageRef).bkg_enqueue priority: 10 # Must be enqueued as a PageRef b/c subclasses aren't recognized by DJ
+      puts "Enqueued #{page_ref.class.to_s} ##{page_ref.id} '#{page_ref.url}' to get status"
+      page_ref.bkg_wait
+      puts "...returned"
+      sentences << "Ran #{page_ref.class.to_s} ##{page_ref.id} '#{page_ref.url}' err_msg #{page_ref.error_message} to get status: now #{page_ref.http_status}"
+    else
+      sentences << "#{page_ref.class.to_s} ##{page_ref.id} '#{page_ref.url}' is okay: status '#{page_ref.status}' http_status = #{page_ref.http_status}, url #{page_ref.url_changed? ? '' : 'not '} changed."
+    end
+    if (page_ref.http_status == 666) &&
+        (match = page_ref.error_message.match(/tried to assert existing url '(.*)'$/)) &&
+        match[1] &&
+        (extant = page_ref.class.find_by_url match[1]) &&
+        (extant.id != page_ref.id)
+      sentences << "Found match PageRef ##{extant.id}"
+      absorb extant
+      page_ref.good!
+    end
+    sentences.join "\n\t"
   end
 
   # Make sure the page_ref has a site
@@ -247,24 +301,24 @@ class PageRefServices
     ([page_ref.url] + page_ref.aliases).each { |old_url|
       if old_url.match(old_ptn)
         new_url = old_url.sub old_ptn, subst
-        puts "Trying to substitute #{new_url} for #{old_url}"
+        puts "Trying to substitute #{new_url} for #{old_url} on PageRef ##{page_ref.id}"
         klass = page_ref.class
         new_page_ref = klass.fetch new_url
+        puts "...got PageRef ##{new_page_ref.id || '<nil>'} '#{new_page_ref.url}' http_status #{new_page_ref.http_status}"
         unless new_page_ref.errors.any?
-          if new_page_ref.id
+          if new_page_ref.id # Existed prior =>
             # Make the old page_ref represent the new URL
-            new_page_ref.aliases += [new_url] unless new_page_ref.url == new_url
             PageRefServices.new(new_page_ref).absorb page_ref
             return new_page_ref
-          elsif extant = klass.find_by(klass.url_query new_page_ref.url) # new_page_ref.url is already found
-            absorb extant
-            return page_ref
+          elsif extant = klass.find_by_url(new_page_ref.url) # new_page_ref.url already exists
+            puts "...returning ##{page_ref.id || '<nil>'} '#{page_ref.url}' http_status #{page_ref.http_status}"
+            epr = PageRefServices.new extant
+            epr.absorb page_ref
+            epr.absorb new_page_ref
+            return extant
           else
-            # Save the old url as an alias. (Presumably not violating uniqueness)
-            page_ref.aliases += [page_ref.url]
-            # Adopt the new_url (which may have been redirected from new_url)
-            page_ref.url = new_page_ref.url
-            absorb new_page_ref
+            absorb new_page_ref, true
+            puts "...returning ##{page_ref.id || '<nil>'} '#{page_ref.url}' http_status #{page_ref.http_status}"
             return page_ref
           end
         end
@@ -273,4 +327,22 @@ class PageRefServices
     nil
   end
 
+  # Check for url collisions between pagerefs of a given type.
+  # This applies to aliases as well as urls, i.e. any given url must map uniquely to one pageref, whether through
+  # the url attribute or via aliases.
+  # This method repairs the trivial case in which a pageref's url also appears in its aliases
+  def self.report_urls type=:recipe
+    klass = "#{type.to_s.capitalize}PageRef".constantize
+    map = {}
+    klass.all.pluck(:id, :url, :aliases).each { |arr|
+      id, url, aliases = *arr
+      if aliases.include? url
+        pr = PageRef.find(id)
+        pr.aliases.delete url
+        pr.save
+      end
+      (aliases + [url]).each { |url| (map[url] ||= []) << id }
+    }
+    map.keep_if { |url, set| set.count > 1 }
+  end
 end
